+++ example/lib/main.dart
@@ -1,38 +1,18 @@
 import 'package:flutter/material.dart';
 
 void main() {
-  runApp(const GuardExampleApp());
+  runApp(const ExampleApp());
 }
 
-class GuardExampleApp extends StatelessWidget {
-  const GuardExampleApp({super.key});
+class ExampleApp extends StatelessWidget {
+  const ExampleApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       home: Scaffold(
-        body: Column(
-          children: [
-            // expect_lint: avoid_native_text_field
-            const TextField(),
-
-            // expect_lint: avoid_native_text_field
-            TextFormField(),
-
-            const CustomTextField(),
-          ],
-        ),
+        body: Padding(padding: EdgeInsets.all(16), child: TextField()),
       ),
     );
-  }
-}
-
-/// Represents the application's approved design-system text field.
-class CustomTextField extends StatelessWidget {
-  const CustomTextField({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const SizedBox();
   }
 }